VideoSummarizer/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.54.0",
    "ffmpeg>=1.4",
    "ffmpeg-python>=0.2.0",
    "openai>=1.86.0",
    "pillow>=11.2.1",
    "pydub>=0.25.1",
    "pypdf2>=3.0.1",
    "reportlab>=4.4.1",
    "streamlit>=1.45.1",
]
